Update infer_with_sliding_window
Add the function infer_with_sliding_window the capability to handle large window size and small image shape. 

If the image is smaller than the window in all dimensions, then no sliding window is done. 

If the image is smaller than the window in one or two dimensions, then the real window size is recorded to create correct size mask_shape and c1